+++ localtv/migrations/0031_create_sitelocation_thumbnails.py
@@ -34,6 +34,8 @@ class Migration:
             sl.logo.open()
             cf = ContentFile(sl.logo.read())
             sl.save_thumbnail_from_file(cf)
+            sl.has_thumbnail = True
+            sl.save()
     
     def backwards(self, orm):
         "Write your backwards migration here"